+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2.5] - 2024-05-26
+- RigidbodyEventVar added to samples
+- EventVarCollection now has more import/export functions
+
 ## [0.2.4] - 2024-05-26
 - bug fixes
 - comments and documentation